World cleanup, Player/Alien decoupling, Factory
Created a factory to handle creating sprites and
to cleanup the World class (previously called Level).

Did away with 'all' group, and replaced with an 'drawable'
and 'updatable' group. To support this, also created an
'interface' called IUpdatable that a sprite object must
implement if it requires an update each frame.

Split up Player class into Player and Alien to support
the possibility of an AI-controlled Alien object.

Added a physics component attribute for some sprites
which require movement vectors or some specific
collision_handling.

Other changes made that should have come in
more than just this commit :)