BaseSprite getter and setter, Factory rename
Added a getter and setter to the BaseSprite class
so that whenever a new image is set for a BaseSprite
object, its rect property is correctly updated to
the correct width and height, while also retaining
its correct position.

Fixed Slime bug that resulted from hit_rect being a
reference to rect by default; solution was to make
a copy upon init, as in other places like Alien.

Moreover, renamed Factory to SpriteFactor for clarity.