Fixed collison bugs and enhanced player jump
For halt_collisions, like between a sprite and
a platform, the edge case was not being considered.
For example, if rect1 had a top y of 100, and rect2
had a bottom y of 2 before the collision, the collision
was detected but not processed correctly; now fixed.

A player may not trigger right and left to "turn" midair
slightly.

Fixed a physics bug concerning how the kinematic equations
were calculated (had a=vt, instead v=at).

Tweaked the Slime mob's movement a little so it would
not get stuck on walls.

Also, the physics component now makes it
clear that it uses the centerx and bottomy
for the position vector.